Cleanup: Rename action handler for consistency
