+++ src/handlers/actions/index.ts
@@ -1,3 +1,3 @@
 export { default as handleMintTokensAction } from './mintTokens';
 export { default as handlePaymentAction } from './payment';
-export { default as handleDomainCreatedAction } from './domainCreated';
+export { default as handleCreateDomainAction } from './createDomain';
